+++ app/build.gradle
@@ -18,7 +18,7 @@ android {
     compileSdkVersion 27
     buildToolsVersion '27.0.2'
     defaultConfig {
-        applicationId 'com.hayaisoftware.launcher'
+        applicationId 'com.anpmech.launcher'
         minSdkVersion 15
         targetSdkVersion 27
         versionCode 10405
