+++ app/src/main/java/com/anpmech/launcher/comparators/RecentOrder.java
@@ -12,16 +12,29 @@
  * limitations under the License.
  */
 
-package com.hayaisoftware.launcher.comparators;
+package com.anpmech.launcher.comparators;
 
-import com.hayaisoftware.launcher.LaunchableActivity;
+import com.anpmech.launcher.LaunchableActivity;
 
 import java.util.Comparator;
 
-public class PinToTop implements Comparator<LaunchableActivity> {
+
+public class RecentOrder implements Comparator<LaunchableActivity> {
 
     @Override
     public int compare(final LaunchableActivity lhs, final LaunchableActivity rhs) {
-        return rhs.getPriority() - lhs.getPriority();
+        final long lhsLaunchTime = lhs.getLaunchTime();
+        final long rhsLaunchTime = rhs.getLaunchTime();
+        final int compare;
+
+        if (lhsLaunchTime > rhsLaunchTime) {
+            compare = -1;
+        } else if (lhsLaunchTime < rhsLaunchTime) {
+            compare = 1;
+        } else {
+            compare = 0;
+        }
+
+        return compare;
     }
 }
