+++ app/src/main/java/com/anpmech/launcher/threading/SimpleTaskConsumerManager.java
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package com.hayaisoftware.launcher.threading;
+package com.anpmech.launcher.threading;
 
 import android.util.Log;
