+++ app/src/main/res/values/strings.xml
@@ -13,8 +13,8 @@
   -->
 
 <resources>
-    <string name="app_name" translatable="false">Hayai Launcher</string>
-    <string name="title_activity_search" translatable="false">Hayai Launcher</string>
+    <string name="app_name" translatable="false">Keikai Launcher</string>
+    <string name="title_activity_search" translatable="false">Keikai Launcher</string>
     <string name="action_settings">Settings</string>
     <string name="application_icon_desc">Application Icon</string>
     <string name="search_icon">Search Icon</string>
@@ -30,7 +30,7 @@
     <string name="action_system_settings">Device settings</string>
     <string name="activity_not_found">Activity not found.</string>
     <string name="action_about">About</string>
-    <string name="about_header">About Hayai Launcher</string>
+    <string name="about_header">About Keikai Launcher</string>
     <string name="about_version_code">Version</string>
     <string name="about_project_website">Project Website</string>
     <string name="about_license">License</string>
